Item checking functionality and UI stabilizations
Border radius values are now pulled from --ion-border-radius
consistently applied across components.

Item card now has its own component and styles.
    - Upon clicking on an item card, it checks the item as packed or,
    unpacked.[1]

[1] This behavior will be changed because of the fact that there will be
store items, unpacked items, and packed items. I will be checking if
the item is a store item or not and if it is I will prompt the user
to add it to the baggage or just mark it as bought (ready).