Baggage filtering, Util function fix, Input checks
Baggages page now actually filters the bags based on the travel selected

While trying to add filtering ı have realised I have been editing the
travels.<baggageUUID> and not <travels.<ActualTravelUUID>>. After
fixing that I realised that the util function was pushing the data
instead of replacing it.

Add item page1 now has a length check for the input field and if the
option "I dont have this item yet" is selected, the checkbox of the
preview card will be disabled.