CurrencySelector added, BaggleDaysLabel updated.
Currency Selector now works and changes the currency setting.
Currrency values displayed in various places.

Days Label updated and now it can also show full date with the remaining
days if the boolean switch is set to true

Storage util function is written to calculate the total price of items
in a baggage.