Fixes in CategoriesBar, Removal of StatusBar.
StatusBar started causing some problems I really don't understand.
I tried fixing it multiple times but it just decided not to work.
So I just removed it for now. I will try to fix it later if I need it

Categories Bar is getting improved with better styles and functionality.
Soon to be able to add categories to a baggage and will let you filter
items by its categories. Hopefully...